fix: refactor app design, fix eslint
