--- src/features/users/user/services/index.ts
@@ -1,2 +1,0 @@
-export * from './user.repository';
-export * from './user.service';
